+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+<a name="0.2.0"></a>
+## [0.2.0](https://github.com/jpcx/node-kraken-api/tree/0.2.0) (2018-06-26)
+
+| __[Changes since 0.1.3](https://github.com/jpcx/node-kraken-api/compare/0.1.3...0.2.0)__ | [Release Notes](https://github.com/jpcx/node-kraken-api/releases/tag/0.2.0) | [README](https://github.com/jpcx/node-kraken-api/tree/0.2.0/README.md) |
+| --- | --- | --- |
+
+| [Source Code (zip)](https://github.com/jpcx/node-kraken-api/archive/0.2.0.zip) | [Source Code (tar.gz)](https://github.com/jpcx/node-kraken-api/archive/0.2.0.tar.gz) |
+| --- | --- |
+
+#### BREAKING CHANGES
+  + __Settings:__ Rate Limiter settings have been reduced to only baseIntvl and minIntvl.
+  + __Sync:__     Sync instance data is no longer non-writable and non-configurable.
+  + __Sync:__     Sync instance time property is now directly below the data property.
+  + __Sync:__     RemoveAllListeners has been removed.
+
+#### Features
+  + __Settings:__ DataFormatter functionality added.
+
+#### Bugfixes
+  + __node-kraken-api:__ SetOTP now returns true.
+  + __Docs:__            API\~Syncing\~EventListener documentation has been updated to reflect new behavior.
+
 <a name="0.1.3"></a>
 ## [0.1.3](https://github.com/jpcx/node-kraken-api/tree/0.1.3) (2018-06-24)